saas/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.74",
    "django>=5.2.8",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
]
prod = [
    "gunicorn>=23.0.0",
]
